Move: StackedBars and StackedAreas go together. I had wrongly put StackedBars with Bars.
