+++ explorer/indicatorExplorer/ExploreView.jsdom.test.tsx
@@ -6,8 +6,8 @@ import xhrMock from "xhr-mock"
 import { ExploreView } from "./ExploreView"
 import { ChartTypeName } from "grapher/core/GrapherConstants"
 import { LineChart } from "grapher/lineCharts/LineChart"
-import { StackedAreaChart } from "grapher/areaCharts/StackedAreaChart"
-import { StackedBarChart } from "grapher/barCharts/StackedBarChart"
+import { StackedAreaChart } from "grapher/stackedCharts/StackedAreaChart"
+import { StackedBarChart } from "grapher/stackedCharts/StackedBarChart"
 import { DiscreteBarChart } from "grapher/barCharts/DiscreteBarChart"
 import { SlopeChart } from "grapher/slopeCharts/SlopeChart"
 import { MapChart } from "grapher/mapCharts/MapChart"
