+++ grapher/chart/ChartTypeMap.ts
@@ -2,9 +2,9 @@ import { ScatterPlotChart } from "grapher/scatterCharts/ScatterPlotChart"
 import { TimeScatterChart } from "grapher/scatterCharts/TimeScatterChart"
 import { SlopeChart } from "grapher/slopeCharts/SlopeChart"
 import { LineChart } from "grapher/lineCharts/LineChart"
-import { StackedAreaChart } from "grapher/areaCharts/StackedAreaChart"
+import { StackedAreaChart } from "grapher/stackedCharts/StackedAreaChart"
 import { DiscreteBarChart } from "grapher/barCharts/DiscreteBarChart"
-import { StackedBarChart } from "grapher/barCharts/StackedBarChart"
+import { StackedBarChart } from "grapher/stackedCharts/StackedBarChart"
 import { ChartTypeName } from "grapher/core/GrapherConstants"
 import { MapChart } from "grapher/mapCharts/MapChart"
