+++ grapher/stackedCharts/AbstractStackedChart.tsx
