+++ grapher/stackedCharts/StackedAreaChart.stories.tsx
