+++ grapher/stackedCharts/StackedAreaChart.test.ts
