+++ grapher/stackedCharts/StackedBarChart.stories.tsx
