+++ grapher/stackedCharts/StackedBarChart.test.ts
