+++ grapher/stackedCharts/StackedBarChart.tsx
