+++ grapher/stackedCharts/StackedConstants.ts
