+++ grapher/stackedCharts/StackedUtils.test.ts
