+++ grapher/stackedCharts/StackedUtils.ts
