+++ utils/test-all-stories.jsdom.test.tsx
@@ -9,7 +9,7 @@ import { mount } from "enzyme"
 
 import * as SwitcherExplorer from "explorer/client/SwitcherExplorer.stories"
 import * as CovidExplorer from "explorer/covidExplorer/CovidExplorer.stories"
-import * as StackedAreaChart from "grapher/areaCharts/StackedAreaChart.stories"
+import * as StackedAreaChart from "grapher/stackedCharts/StackedAreaChart.stories"
 import * as DiscreteBarChart from "grapher/barCharts/DiscreteBarChart.stories"
 import * as CaptionedChart from "grapher/captionedChart/CaptionedChart.stories"
 import * as NoDataModal from "grapher/noDataModal/NoDataModal.stories"
@@ -28,7 +28,7 @@ import * as ScatterPlot from "grapher/scatterCharts/ScatterPlotChart.stories"
 import * as SlopeChart from "grapher/slopeCharts/SlopeChart.stories"
 import * as TimelineComponent from "grapher/timeline/TimelineComponent.stories"
 import * as Feedback from "site/client/Feedback.stories"
-import * as StackedBarChart from "grapher/barCharts/StackedBarChart.stories"
+import * as StackedBarChart from "grapher/stackedCharts/StackedBarChart.stories"
 import * as DownloadTab from "grapher/downloadTab/DownloadTab.stories"
 import * as LineLegend from "grapher/lineLegend/LineLegend.stories"
 import * as MapChart from "grapher/mapCharts/MapChart.stories"
